reorganise the components
